// src/shared/enums.ts
export enum THEMES {
  LIGHT = "light",
  DARK = "dark",
}

export enum ImageSize {
  DEFAULT = "default",
  FUll = "full",
  SMALL = "small",
  MEDIUM = "medium",
  XS = 'xs'
}

export enum NavbarType {
  DEFAULT = "default",
  CENTERED = "centered",
}

export enum TextAlign {
  LEFT = "left",
  RIGHT = "right",
  CENTER = "center",
  JUSTIFY = "justify",
  NONE = "unset"
}

export enum LogoType {
  TEXT = "text",
  IMAGE = "image",
}

export enum ListType {
  disc = "list-disc",
  number = "list-decimal",
  none = "list-none"
}